+++ src/share/vm/gc_implementation/shenandoah/shenandoahMonitoringSupport.cpp
@@ -32,7 +32,7 @@
 class DummyGenerationCounters : public GenerationCounters {
 public:
   DummyGenerationCounters():
-    GenerationCounters("dummy", 0, 3, MinObjAlignmentInBytes * 3, MinObjAlignmentInBytes * 3, (size_t) 0) {
+    GenerationCounters("dummy", 0, 3, (size_t)(MinObjAlignmentInBytes * 3), (size_t)(MinObjAlignmentInBytes * 3), (size_t) 0) {
     _current_size->set_value(0);
     // Nothing to do.
   }
